Simplify code, add protocol-agnostic FrameSink, fix compiler warnings
Deduplicate and simplify shared algorithms: extract dechirp_argmax() and
dewhiten() to utilities/tx_chain, upgrade hamming_encode_frame() to std::span,
remove dead double_mod(), simplify add_crc() to call crc_to_nibbles().
Pre-allocate 14 heap buffers in BurstDetector as class members.

Add protocol-agnostic FrameSink with CBOR output schema (PHY-only fields),
minimal CBOR encoder, and extracted MeshCore v1 parser. ConsoleSink becomes
a thin wrapper adding MeshCore protocol decoding on top of FrameSink.

Fix all compiler warnings in production headers and test files: int-to-size_t
loop indexing, uint16_t narrowing from shift/OR expressions, unused variables,
and mod() signed/unsigned cast safety in interleaving.

+++ blocks/include/gnuradio-4.0/lora/algorithm/crc.hpp
@@ -63,7 +63,7 @@ namespace gr::lora {
 /// payload_with_crc contains payload_len bytes of data followed by 2 CRC bytes (reassembled).
 /// Returns true if CRC matches.
 [[nodiscard]] inline bool lora_verify_crc(std::span<const uint8_t> payload, uint8_t crc_lo, uint8_t crc_hi) {
-    uint16_t received_crc = static_cast<uint16_t>(crc_lo) | (static_cast<uint16_t>(crc_hi) << 8);
+    uint16_t received_crc = static_cast<uint16_t>(static_cast<unsigned>(crc_lo) | (static_cast<unsigned>(crc_hi) << 8));
     uint16_t computed     = lora_payload_crc(payload);
     return received_crc == computed;
 }
